+++ src/bpInfo.ts
@@ -1,1 +1,1 @@
-export const bpInfo = {version: 'v1.2.8'};
+export const bpInfo = {version: 'v1.2.9'};